# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Manipulação de livros e suas dependências ",
    "version" : "v1",
    "title" : "Livraria Virtual"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/IKokiri/Livrariav/v1",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/livros/{carrinhoId}/pedido" : {
      "post" : {
        "tags" : [ "pedido" ],
        "summary" : "Adiciona um carrinho ao pedido",
        "operationId" : "addCarrinhoPedido",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carrinhoId",
          "in" : "path",
          "description" : "ID do do carrinho ao ser adicionado ao pedido",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do carrinho que será adicionado ao pedido",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Carrinho"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      },
      "delete" : {
        "tags" : [ "pedido" ],
        "summary" : "Remoção de um carrinho do pedido",
        "description" : "Remove um carrinho de um pedido",
        "operationId" : "deleteLivroCarrinhoPedido",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carrinhoId",
          "in" : "path",
          "description" : "identificador do carrinho que será removido do pedido",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "400" : {
            "description" : "carrinho inválido"
          },
          "404" : {
            "description" : "carrinho não está no pedido encontrado"
          }
        }
      }
    },
    "/livros/pedido/{pedidoId}/acompanhamento" : {
      "get" : {
        "tags" : [ "pedido" ],
        "summary" : "Busca o status de um pedido",
        "description" : "faz a busca no banco do status de um pedido",
        "operationId" : "getStatusPedido",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "pedidoId",
          "in" : "path",
          "description" : "ID do pedido",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "seccess",
            "schema" : {
              "$ref" : "#/definitions/Pedido"
            }
          },
          "400" : {
            "description" : "Id inválido"
          },
          "404" : {
            "description" : "Não Encontrado"
          }
        }
      }
    }
  },
  "definitions" : {
    "Livro" : {
      "type" : "object",
      "properties" : {
        "isbn" : {
          "type" : "number"
        },
        "titulo" : {
          "type" : "string"
        },
        "autor" : {
          "$ref" : "#/definitions/Autor"
        },
        "genero" : {
          "$ref" : "#/definitions/Genero"
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "preco" : {
          "type" : "number"
        }
      },
      "xml" : {
        "name" : "Livro"
      },
      "example" : {
        "preco" : 5.962133916683182377482808078639209270477294921875,
        "data" : "2000-01-23T04:56:07.000+00:00",
        "isbn" : 0.80082819046101150206595775671303272247314453125,
        "genero" : {
          "genero" : "genero",
          "id" : 1
        },
        "titulo" : "titulo",
        "autor" : {
          "nome" : "nome",
          "id" : 6
        }
      }
    },
    "Autor" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "nome" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Autor"
      },
      "example" : {
        "nome" : "nome",
        "id" : 6
      }
    },
    "Genero" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "genero" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Genero"
      },
      "example" : {
        "genero" : "genero",
        "id" : 1
      }
    },
...
    "Carrinho" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "number"
        },
        "livro" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Livro"
          }
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "total" : {
          "type" : "number"
        }
      },
      "xml" : {
        "name" : "Carrinho"
      },
      "example" : {
        "total" : 6.02745618307040320615897144307382404804229736328125,
        "data" : "2000-01-23T04:56:07.000+00:00",
        "livro" : [ {
          "preco" : 5.962133916683182377482808078639209270477294921875,
          "data" : "2000-01-23T04:56:07.000+00:00",
          "isbn" : 0.80082819046101150206595775671303272247314453125,
          "genero" : {
            "genero" : "genero",
            "id" : 1
          },
          "titulo" : "titulo",
          "autor" : {
            "nome" : "nome",
            "id" : 6
          }
        }, {
          "preco" : 5.962133916683182377482808078639209270477294921875,
          "data" : "2000-01-23T04:56:07.000+00:00",
          "isbn" : 0.80082819046101150206595775671303272247314453125,
          "genero" : {
            "genero" : "genero",
            "id" : 1
          },
          "titulo" : "titulo",
          "autor" : {
            "nome" : "nome",
            "id" : 6
          }
        } ],
        "id" : 0.80082819046101150206595775671303272247314453125
      }
    },
    "Pedido" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "number"
        },
        "livro" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Livro"
          }
        },
        "status" : {
          "type" : "string"
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "total" : {
          "type" : "number"
        }
      },
      "xml" : {
        "name" : "Carrinho"
      },
      "example" : {
        "total" : 6.02745618307040320615897144307382404804229736328125,
        "data" : "2000-01-23T04:56:07.000+00:00",
        "livro" : [ {
          "preco" : 5.962133916683182377482808078639209270477294921875,
          "data" : "2000-01-23T04:56:07.000+00:00",
          "isbn" : 0.80082819046101150206595775671303272247314453125,
          "genero" : {
            "genero" : "genero",
            "id" : 1
          },
          "titulo" : "titulo",
          "autor" : {
            "nome" : "nome",
            "id" : 6
          }
        }, {
          "preco" : 5.962133916683182377482808078639209270477294921875,
          "data" : "2000-01-23T04:56:07.000+00:00",
          "isbn" : 0.80082819046101150206595775671303272247314453125,
          "genero" : {
            "genero" : "genero",
            "id" : 1
          },
          "titulo" : "titulo",
          "autor" : {
            "nome" : "nome",
            "id" : 6
          }
        } ],
        "id" : 0.80082819046101150206595775671303272247314453125,
        "status" : "status"
      }
    }
  }
}